AutoControllerで待機する閉塞数を設定出るようにした
・StayDistanceの追加
・Untenクラスで，StayDistanceとVehicle.Distanceを比較して，Phaseの進行を待機するように変更

+++ ServerUtility/TusSolution/AutoController/AutoController/Deserialzied/DeserializedVehicle.cs
@@ -23,7 +23,7 @@ namespace Tus.AutoController.Deserialized
         public string[] ReallocatableBlockNames { get; set; }
         public string ShownName { get; set; }
         public virtual float Speed { get; set; }
-        public int StopThreshold { get; set; }
+        public float StopThreshold { get; set; }
         public int VehicleID { get; set; }
     }
     public class DeserializedBlock
